+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## UNRELEASED
 * Rename NoEmptyIfDefinedPlugin to RedundantExistenceChecksPlugin (Daimona Eaytoy)
 * Add `isset()` checks to RedundantExistenceChecksPlugin (Daimona Eaytoy)
+* Upgrade phan to 5.4.5 and mediawiki/phan-taint-check-plugin to 6.1.0 (James D. Forrester)
 
 ## 0.14.0 / 2024-02-03
 * Add plugin to disallow use of `new Exception` (Daimona Eaytoy)